requirements.txt
python-dotenv
requests
pandas
numpy
pymongo
certifi
pymongo[srv]==3.6

-e .